Create category done .

+++ category-service/src/main/java/com/mehmetpekdemir/categoryservice/common/rest/MetaResponse.java
@@ -13,9 +13,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 class MetaResponse {
 
-    private int limit;
-    private int currentPage;
-    private int total;
-    private int totalPage;
+    private Integer limit;
+    private Integer currentPage;
+    private Long total;
+    private Integer totalPage;
 
 }
